+++ src/sentry/static/sentry/app/data/forms/projectDebugFiles.jsx
@@ -3,20 +3,11 @@ import React from 'react';
 
 import {t} from 'app/locale';
 import {openDebugFileSourceModal} from 'app/actionCreators/modal';
+import {DEBUG_SOURCE_TYPES} from 'app/data/debugFileSources';
 import TextBlock from 'app/views/settings/components/text/textBlock';
 
 // Export route to make these forms searchable by label/help
 export const route = '/settings/:orgId/projects/:projectId/debug-symbols/';
-
-export const sourceNames = {
-  gcs: t('Google Cloud Storage'),
-  http: t('SymbolServer (HTTP)'),
-  s3: t('Amazon S3'),
-};
-
-export function getSourceName(type) {
-  return sourceNames[type] || t('Unknown');
-}
 
 function flattenKeys(obj) {
   const result = {};
@@ -64,17 +55,17 @@ export const fields = {
       items: [
         {
           value: 's3',
-          label: sourceNames.s3,
+          label: t(DEBUG_SOURCE_TYPES.s3),
           searchKey: t('aws amazon s3 bucket'),
         },
         {
           value: 'gcs',
-          label: sourceNames.gcs,
+          label: t(DEBUG_SOURCE_TYPES.gcs),
           searchKey: t('gcs google cloud storage bucket'),
         },
         {
           value: 'http',
-          label: sourceNames.http,
+          label: t(DEBUG_SOURCE_TYPES.http),
           searchKey: t('http symbol server ssqp symstore symsrv'),
         },
       ],
@@ -84,11 +75,7 @@ export const fields = {
     setValue: raw => (JSON.parse(raw || null) || []).map(flattenKeys),
 
     renderItem(item) {
-      if (item.name) {
-        return item.name;
-      } else {
-        return <em>{t('<Unnamed Repository>')}</em>;
-      }
+      return item.name || <em>{t('<Unnamed Repository>')}</em>;
     },
 
     onAddItem(item, addItem) {
@@ -110,7 +97,7 @@ export const fields = {
       title: t('Remove Repository?'),
       confirmText: t('Remove Repository'),
       message: (
-        <>
+        <React.Fragment>
           <TextBlock>
             <strong>
               {t('Removing this repository applies instantly to new events.')}
@@ -121,7 +108,7 @@ export const fields = {
               'Debug files from this repository will not be used to symbolicate future events. This may create new issues and alert members in your organization.'
             )}
           </TextBlock>
-        </>
+        </React.Fragment>
       ),
     },
   },
